+++ src/tools/testing/index.ts
@@ -13,6 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export * from './sf-get-username.js';
-export * from './sf-resume.js';
+export * from './sf-test-apex.js';
+export * from './sf-test-agents.js';